refactor: to separate components
